+++ beeref/commands.py
@@ -325,7 +325,7 @@ class ChangeText(QtGui.QUndoCommand):
 
 
 class ChangeOpacity(QtGui.QUndoCommand):
-    """Change Opacity."""
+    """Change opacity on images."""
 
     def __init__(self, items, opacity, ignore_first_redo=False):
         super().__init__('Change Opacity')
@@ -345,3 +345,21 @@ class ChangeOpacity(QtGui.QUndoCommand):
     def undo(self):
         for item, opacity in zip(self.items, self.old_opacities):
             item.setOpacity(opacity)
+
+
+class ToggleGrayscale(QtGui.QUndoCommand):
+    """Toggle grayscale mode on images."""
+
+    def __init__(self, items, grayscale):
+        super().__init__('Toggle Grayscale')
+        self.items = list(filter(lambda item: item.is_image, items))
+        self.grayscale = grayscale
+        self.old_grayscales = [item.grayscale for item in items]
+
+    def redo(self):
+        for item in self.items:
+            item.grayscale = self.grayscale
+
+    def undo(self):
+        for item, grayscale in zip(self.items, self.old_grayscales):
+            item.grayscale = grayscale