+++ beeref/view.py
@@ -287,6 +287,14 @@ class BeeGraphicsView(MainControlsMixin,
             self.scene.selectedItems(user_only=True)))
         widgets.ChangeOpacityDialog(self, images, self.undo_stack)
 
+    def on_action_grayscale(self, checked):
+        images = list(filter(
+            lambda item: item.is_image,
+            self.scene.selectedItems(user_only=True)))
+        if images:
+            self.undo_stack.push(
+                commands.ToggleGrayscale(images, checked))
+
     def on_action_crop(self):
         self.scene.crop_items()