+++ tests/fileio/test_sql.py
@@ -235,6 +235,7 @@ def test_sqliteio_write_inserts_new_pixmap_item_png(tmpfile, view):
     item.setRotation(33)
     item.do_flip()
     item.crop = QtCore.QRectF(5, 5, 100, 80)
+    item.grayscale = True
     item.pixmap_to_bytes = MagicMock(return_value=(b'abc', 'png'))
     io = SQLiteIO(tmpfile, view.scene, create_new=True)
     io.write()
@@ -255,6 +256,7 @@ def test_sqliteio_write_inserts_new_pixmap_item_png(tmpfile, view):
         'filename': 'bee.jpg',
         'crop': [5, 5, 100, 80],
         'opacity': 0.66,
+        'grayscale': True,
     }
     assert result[7] == 'pixmap'
     assert result[8] == b'abc'
@@ -346,6 +348,7 @@ def test_sqliteio_write_updates_existing_pixmap_item(tmpfile, view):
     item.setOpacity(0.75)
     item.do_flip()
     item.crop = QtCore.QRectF(1, 2, 30, 40)
+    item.grayscale = True
     item.filename = 'new.png'
     item.pixmap_to_bytes.return_value = b'updated'
     io.create_new = False
@@ -366,6 +369,7 @@ def test_sqliteio_write_updates_existing_pixmap_item(tmpfile, view):
         'filename': 'new.png',
         'crop': [1, 2, 30, 40],
         'opacity': 0.75,
+        'grayscale': True,
     }
     assert result[7] == b'abc'
 
@@ -502,6 +506,7 @@ def test_sqliteio_read_reads_readonly_pixmap_item(tmpfile, view, imgdata3x3):
     assert item.height == 3
     assert item.crop == QtCore.QRectF(0, 0, 3, 3)
     assert item.opacity() == 1
+    assert item.grayscale is False
     assert view.scene.items_to_add.empty() is True